Add Project.toml and Manifest.toml, rewrite WIP
