+++ Project.toml
@@ -2,5 +2,10 @@ name = "DiscriminantAnalysis"
 uuid = "2091bfbe-2937-51d6-a893-7459a14235e5"
 version = "0.1.0"
 
+[deps]
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
+
 [compat]
-julia = "0.4.0"
+julia = "1"
+StatsBase = "0.30, 0.31, 0.32"
