+++ src/DiscriminantAnalysis.jl
@@ -1,24 +1,7 @@
 module DiscriminantAnalysis
+    using StatsBase, LinearAlgebra
 
-    import Base: 
-        LinAlg.BlasReal, 
-        show,
-        AbstractArray,
-        size,
-        linearindexing,
-        getindex, 
-        length, 
-        convert
-
-    export
-        lda,
-        cda,
-        qda,
-        discriminants,
-        classify
+    abstract type DiscriminantModel <: StatsBase.StatisticalModel end
 
     include("common.jl")
-    include("lda.jl")
-    include("qda.jl")
-
-end # module
+end
